Create folder and read me.md file
Creating folder rather than just an .md file so that there is a home for other files and such related to this buy. That way we can put things in there like helpful background info - slides from 6/6/2019, org chart, current tech stack info. That said, I'm envisioning the draft Statements of Objectives and the draft RFQ to be in a markdown file. So far I've drafted some background for the reader when they first come to this repo.