Create soo.md
1. Taking a stab at the Problem Statement

2. The single product vision in the .md file is an attempt to put together a number of elements from the following problem statements that were provided by workshopping tables on 6/6/2019:  

Empathy & education around the history of slavery and freedom in America from 1400 to 1877 needs to be increased through a digital experience. The solution should evolve to encompass and deliver the fully collected, curated body of knowledge available at the NMAAHC  

We believe that bringing the emotional experience of slavery and freedom exhibition online audience and for those who are seeking to engage African American culture will allow people to connect with African American history and culture to understand how it affects our world today, we will know this is true when we see “American” story becomes the center of the African American story.  

How might we elevate and engage the narrative we want to help visitors both online and in person to better connect with and understand the history of slavery and freedom to help us make better decisions today.  

How might we engage partners in healthy competition to create a digital experience in one years’ time that conveys the emotional human journey from slavery and freedom that promotes discovery in an inclusive way which is also informed by research and prototyping.  (this is too much focused on the one-year thing. I think that. I think the problem statement should live on beyond one year)