Create rfq.md
Adding rfq.md which will be a home for a draft Request for Quote/Proposal. This is not really complete enough to constitute a first draft, but I'm approaching this in an iterative way. You'll notice questions to encourage feedback from the vendors. You'll also notice I includes several potential acquisition strategies, which I especially would like to get feedback from the vendors on.